+++ common/src/main/java/com/verdantartifice/primalmagick/common/crafting/IWandTransform.java
@@ -18,7 +18,27 @@ public interface IWandTransform {
      * @param pos the position of the thing to be transformed
      * @return true if the player may proceed with the transform, false otherwise
      */
-    public boolean isValid(Level world, Player player, BlockPos pos);
+    boolean isValid(Level world, Player player, BlockPos pos);
+
+    /**
+     * Determine whether this transformation is being on run something similar, but not exactly the same as, the
+     * intended target and should show a warning to the player.
+     *
+     * @param world the world in which the transform is taking place
+     * @param player the player attempting to perform the transformation
+     * @param pos the position of the thing to be transformed
+     * @return true if the player should be shown a similarity warning for the transform, false otherwise
+     */
+    boolean isSimilar(Level world, Player player, BlockPos pos);
+
+    /**
+     * Show the pre-defined similarity warning to the given player.
+     *
+     * @param world the world in which the transform is taking place
+     * @param player the player attempting to perform the transformation
+     * @param pos the position of the thing to be transformed
+     */
+    void showSimilarityWarning(Level world, Player player, BlockPos pos);
     
     /**
      * Execute the defined transformation.
@@ -27,5 +47,5 @@ public interface IWandTransform {
      * @param player the player attempting to perform the transformation
      * @param pos the position of the thing to be transformed
      */
-    public void execute(Level world, Player player, BlockPos pos);
+    void execute(Level world, Player player, BlockPos pos);
 }